+++ changelog.md
@@ -1,4 +1,10 @@
 # 0.2.2
 
 - Implement basic user operations: read, read many, update
+- You can update custom fields in:
+  - accounts
+  - issues
+  - leads
+  - opportunities
+  - users